cleaning up signed/unsigned comparison warnings in the rest of Teem
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2469 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/bane/clip.c
@@ -91,7 +91,7 @@ _baneClipAnswer_TopN(int *countP, Nrrd *hvol, double *clipParm) {
   hits = copy->data;
   num = nrrdElementNumber(copy);
   qsort(hits, num, sizeof(int), nrrdValCompare[nrrdTypeInt]);
-  tmp = AIR_CLAMP(0, (int)clipParm[0], num-1);
+  tmp = AIR_CLAMP(0, (int)clipParm[0], (int)num-1);
   *countP = hits[num-tmp-1];
   nrrdNuke(copy);
